+++ packages/opencode/src/server/instance/config.ts
