+++ packages/opencode/src/server/instance/file.ts
