+++ packages/opencode/src/server/instance/global.ts
