+++ packages/opencode/src/server/instance/mcp.ts
