+++ packages/opencode/src/server/instance/permission.ts
