+++ packages/opencode/src/server/instance/project.ts
