+++ packages/opencode/src/server/instance/provider.ts
