+++ packages/opencode/src/server/instance/pty.ts
