+++ packages/opencode/src/server/instance/session.ts
