+++ packages/opencode/src/server/instance/tui.ts
