+++ packages/opencode/src/server/instance/workspace.ts
